+++ JRestaurantSystem/JRestaurantSystem.csproj.user
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Compile Update="Form1.cs">
+    <Compile Update="RestaurantMap.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Update="TableView.cs">
